requirements.txt
discord.py
instagrapi
deep-translator
python-dotenv
PyNaCl